+++ src/audiovideocontroller/AudioVideoControllerState.ts
@@ -29,17 +29,23 @@ import VideoTileController from '../videotilecontroller/VideoTileController';
 import VideoUplinkBandwidthPolicy from '../videouplinkbandwidthpolicy/VideoUplinkBandwidthPolicy';
 import VolumeIndicatorAdapter from '../volumeindicatoradapter/VolumeIndicatorAdapter';
 
+export const DEFAULT_VIDEO_SUBSCRIPTION_LIMIT = 25;
+
 /**
- * [[AudioVideoControllerState]] includes the compute resources shared by [[Task]].
+ * [[AudioVideoControllerState]] includes the compute resources shared by [[DefaultAudioVideoController]] and any running [[Task]].
+ *
+ * **Note**: Any additions to this class need to consider whether they need to be reset in `resetConnectionSpecificState`, `CleanStoppedSessionTask`, or
+ * `CleanRestartedSessionTask`, e.g. if it is relies on backend state and will go stale across reconnections.  Failing
+ * to reset state may lead to unexpected behavior.
  */
 export default class AudioVideoControllerState {
   logger: Logger | null = null;
 
   browserBehavior: ExtendedBrowserBehavior | null = null;
 
-  signalingClient: SignalingClient | null = null;
-
   meetingSessionConfiguration: MeetingSessionConfiguration | null = null;
+
+  signalingClient: SignalingClient | null = null;
 
   peer: RTCPeerConnection | null = null;
 
@@ -55,11 +61,11 @@ export default class AudioVideoControllerState {
 
   mediaStreamBroker: MediaStreamBroker | null = null;
 
-  audioMixController: AudioMixController | null = null;
-
   activeAudioInput: MediaStream | undefined = undefined;
 
   activeVideoInput: MediaStream | undefined = undefined;
+
+  audioMixController: AudioMixController | null = null;
 
   transceiverController: TransceiverController | null = null;
 
@@ -107,7 +113,7 @@ export default class AudioVideoControllerState {
   // This value is set in the `JoinAndReceiveIndexTask` when we process the `SdkJoinAckFrame`
   // and is used in the `ReceiveVideoStreamIndexTask` to limit the total number of streams
   // that we include in the `videosToReceive`.
-  videoSubscriptionLimit: number = 25;
+  videoSubscriptionLimit: number = DEFAULT_VIDEO_SUBSCRIPTION_LIMIT;
 
   // The previous SDP answer will be used as a dictionary to seed the compression library
   // during decompressing the compressed SDP answer.
@@ -121,7 +127,7 @@ export default class AudioVideoControllerState {
 
   // Calculated as the highest priority available codec set in the (possibly munged) SDP answer
   // that is provide to the peer connection, which will be what is sent.
-  currentVideoSendCodec: VideoCodecCapability;
+  currentVideoSendCodec: VideoCodecCapability | undefined = undefined;
 
   // Intersection of `videoSendCodecPreferences` and the supported receive codecs of
   // all the other clients in the meeting.
@@ -137,10 +143,13 @@ export default class AudioVideoControllerState {
 
   connectionMonitor: ConnectionMonitor | null = null;
 
+  // This state is deprecated and unused.
   videoInputAttachedTimestampMs: number = 0;
 
+  // This state is deprecated and unused.
   audioDeviceInformation: { [id: string]: string } = {};
 
+  // This state is deprecated and unused.
   videoDeviceInformation: { [id: string]: string } = {};
 
   enableSimulcast: boolean = false;
@@ -162,4 +171,65 @@ export default class AudioVideoControllerState {
   maxVideoTileCount: number = 0;
 
   startTimeMs: number | null = null;
+
+  /*
+   * Reset state corresponding to state that is dependent on a individual connection
+   * and may not be valid for others, e.g. on a reconnection.
+   */
+  resetConnectionSpecificState(): void {
+    // For auditing reasons, we will comment on the state that we do not touch here. Note that `DefaultAudioVideoController.actionConnect`
+    // also resets certain state, some to cached members:
+    // Reset to empty/null/new state: `browserBehavior`, `transceiverController`, `volumeIndicatorAdapter`, `enableSimulcast`
+    //                                `signalingOpenDurationMs`, `iceGatheringDurationMs`, `startAudioVideoTimestamp`, `attendeePresenceDurationMs`
+    //                                `meetingStartDurationMs`, `startTimeMs`, `lastKnownVideoAvailability`, `videoCaptureAndEncodeParameter`, `videosToReceive`
+    //                                `videosPaused`, `videoStreamIndex`, `statsCollector`, `connectionMonitor`
+    // Reset to existing/cached values: `logger`, `meetingSessionConfiguration`, `realtimeController`, `mediaStreamBroker`,
+    //                                  `audioMixController`, `reconnectController, `audioProfile`, `eventController`
+
+    // `signalingClient` can be reused from a failed/disconnected state.
+
+    if (this.peer) {
+      this.peer.close();
+    }
+    this.peer = null;
+    this.previousSdpOffer = null;
+    this.sdpOfferInit = null;
+
+    // `audioVideoController` members should either be reusable, or moved to `AudioVideoControllerState` and
+    // cleaned up here.
+
+    // We don't want to mutate `videoTileController` as most video tiles will still be there on reconnect. We can remove tiles on the
+    // first index we receive if they no longer exist
+
+    // `mediaStreamBroker`, `activeAudioInput`, and `activeVideoInput` are cleaned up seperately in `DefaultAudioVideoController.cleanUpMediaStreamsAfterStop`
+    // but only on `stop` or non-reconnectable failures. They are also set to cached `DefaultAudioVideoController` members on restart.
+
+    if (this.transceiverController !== undefined) {
+      this.transceiverController.reset();
+    }
+    this.indexFrame = null;
+    this.iceCandidates = [];
+    this.iceCandidateHandler = null;
+    this.sdpAnswer = null;
+    this.turnCredentials = null;
+
+    this.videoDownlinkBandwidthPolicy.reset();
+    if (this.videoUplinkBandwidthPolicy.reset) {
+      this.videoUplinkBandwidthPolicy.reset();
+    }
+
+    this.lastVideosToReceive = null;
+    this.videoSubscriptions = null;
+    this.videoSubscriptionLimit = DEFAULT_VIDEO_SUBSCRIPTION_LIMIT;
+    this.previousSdpAnswerAsString = '';
+    this.serverSupportsCompression = false;
+
+    // `videoSendCodecPreferences` is set by builder and needs to stay consistent.
+
+    this.currentVideoSendCodec = undefined;
+    this.meetingSupportedVideoSendCodecPreferences = undefined;
+    this.videoDuplexMode = null;
+
+    // `poorConnectionCount`and `maxVideoTileCount` is intentionally not set to 0 across reconnections.
+  }
 }